feat(migrations): update evm_tokens schema to use BIGINT for chain_id and implement migration scripts

+++ src/schema.rs
@@ -3,7 +3,7 @@
 diesel::table! {
     evm_tokens (id) {
         id -> Text,
-        chain_id -> Integer,
+        chain_id -> BigInt,
         address -> Text,
         symbol -> Text,
         decimals -> Integer,
